// src/app/models/user.model.ts
export interface User {
    id: string;
    firstname: string;
    lastname: string;
    username: string;
    email: string;
    img_profile: any;
    role_id: string;
    master: string;
    status: boolean;
    password: string;
    remember_token: string;
    api_token: string;
    created_at: string;
    updated_at: string;
    deleted_at: string;
}